prepare for migration to 15, update links to use legacy and rename image components
