+++ components/sections/HeroWide.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import Image from "next/future/image";
+import Image from "next/image";
 import classes from "../../styles/HeroWide.module.css";
 
 interface HeroProps {
