+++ components/ui/ImageModal.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import Button from 'react-bootstrap/Button';
 import { saveAs } from 'file-saver';
 import Modal from 'react-bootstrap/Modal';
-import Image, { StaticImageData } from 'next/image';
+import Image, { StaticImageData } from "next/legacy/image";
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCloudArrowDown } from '@fortawesome/free-solid-svg-icons';
 import styles from '../../styles/ImageModal.module.css';
