+++ next.config.js
@@ -1,16 +1,15 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-  swcMinify: true,
   images: {
     remotePatterns: [
       {
-        protocol: 'https',
-        hostname: 'images.ctfassets.net',
+        protocol: "https",
+        hostname: "images.ctfassets.net",
       },
       {
-        protocol: 'https',
-        hostname: 'downloads.ctfassets.net',
+        protocol: "https",
+        hostname: "downloads.ctfassets.net",
       },
     ],
   },
