+++ package.json
@@ -20,10 +20,10 @@
     "contentful": "^9.2.5",
     "file-saver": "^2.0.5",
     "markdown-to-jsx": "^7.1.7",
-    "next": "12.3.1",
-    "react": "18.2.0",
+    "next": "^15.0.1",
+    "react": "^18.3.1",
     "react-bootstrap": "^2.5.0",
-    "react-dom": "18.2.0",
+    "react-dom": "^18.3.1",
     "sharp": "^0.31.1"
   },
   "devDependencies": {
@@ -38,7 +38,7 @@
     "eslint-config-airbnb": "^19.0.4",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-config-airbnb-typescript": "^17.0.0",
-    "eslint-config-next": "^12.3.1",
+    "eslint-config-next": "^15.0.1",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-jsx-a11y": "^6.6.1",
